requirements.txt
python-dotenv==1.1.1
Requests==2.32.4
streamlit==1.37.0
requests>=2.31.0
python-dotenv>=1.0.0
